remove redundant override createTable
