+++ sql/catalyst/src/test/scala/org/apache/spark/sql/connector/catalog/InMemoryTableCatalog.scala
@@ -84,7 +84,8 @@ class BasicInMemoryTableCatalog extends TableCatalog {
     invalidatedTables.add(ident)
   }
 
-  // TODO: remove it when no tests calling this deprecated method.
+  // TODO: remove it when the deprecated method `createTable(..., StructType, ...)`
+  //  in the parent interface `TableCatalog` is removed
   override def createTable(
       ident: Identifier,
       schema: StructType,
